# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Authentication API",
        "description": "API endpoints for user authentication",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/stats": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "Get all stats",
                "description": "List all stats",
                "operationId": "0a506e816973783ad4c275ef4729b391",
                "responses": {
                    "200": {
                        "description": "List of stats",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "player_id": {
                                                "type": "integer"
                                            },
                                            "game_id": {
                                                "type": "integer"
                                            },
                                            "points": {
                                                "type": "integer"
                                            },
                                            "rebounds": {
                                                "type": "integer"
                                            },
                                            "assists": {
                                                "type": "integer"
                                            },
                                            "player": {
                                                "type": "object"
                                            },
                                            "game": {
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Stats"
                ],
                "summary": "Create a new stat",
                "description": "Create a new stat",
                "operationId": "0f7ee486549de7a5fe9ac24d7cfc0d5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_id",
                                    "game_id"
                                ],
                                "properties": {
                                    "player_id": {
                                        "type": "integer"
                                    },
                                    "game_id": {
                                        "type": "integer"
                                    },
                                    "seconds_played": {
                                        "type": "integer"
                                    },
                                    "points": {
                                        "type": "integer"
                                    },
                                    "field_goals_made": {
                                        "type": "integer"
                                    },
                                    "field_goals_attempted": {
                                        "type": "integer"
                                    },
                                    "three_point_field_goals_made": {
                                        "type": "integer"
                                    },
                                    "three_point_field_goals_attempted": {
                                        "type": "integer"
                                    },
                                    "two_point_field_goals_made": {
                                        "type": "integer"
                                    },
                                    "two_point_field_goals_attempted": {
                                        "type": "integer"
                                    },
                                    "free_throws_made": {
                                        "type": "integer"
                                    },
                                    "free_throws_attempted": {
                                        "type": "integer"
                                    },
                                    "offensive_rebounds": {
                                        "type": "integer"
                                    },
                                    "defensive_rebounds": {
                                        "type": "integer"
                                    },
                                    "total_rebounds": {
                                        "type": "integer"
                                    },
                                    "assists": {
                                        "type": "integer"
                                    },
                                    "turnovers": {
                                        "type": "integer"
                                    },
                                    "steals": {
                                        "type": "integer"
                                    },
                                    "blocks": {
                                        "type": "integer"
                                    },
                                    "personal_fouls": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Stat created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "stat": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stats/{id}": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "Get stat details",
                "description": "Get a specific stat",
                "operationId": "9ad12eeeb82fa3017a2b5fdbbb6d24ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Stat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stat details"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Stats"
                ],
                "summary": "Update stat details",
                "description": "Update a stat",
                "operationId": "8cbf13234c22a3b19cdc535ad6019a18",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Stat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_id",
                                    "game_id"
                                ],
                                "properties": {
                                    "player_id": {
                                        "type": "integer"
                                    },
                                    "game_id": {
                                        "type": "integer"
                                    },
                                    "seconds_played": {
                                        "type": "integer"
                                    },
                                    "points": {
                                        "type": "integer"
                                    },
                                    "field_goals_made": {
                                        "type": "integer"
                                    },
                                    "field_goals_attempted": {
                                        "type": "integer"
                                    },
                                    "three_point_field_goals_made": {
                                        "type": "integer"
                                    },
                                    "three_point_field_goals_attempted": {
                                        "type": "integer"
                                    },
                                    "two_point_field_goals_made": {
                                        "type": "integer"
                                    },
                                    "two_point_field_goals_attempted": {
                                        "type": "integer"
                                    },
                                    "free_throws_made": {
                                        "type": "integer"
                                    },
                                    "free_throws_attempted": {
                                        "type": "integer"
                                    },
                                    "offensive_rebounds": {
                                        "type": "integer"
                                    },
                                    "defensive_rebounds": {
                                        "type": "integer"
                                    },
                                    "total_rebounds": {
                                        "type": "integer"
                                    },
                                    "assists": {
                                        "type": "integer"
                                    },
                                    "turnovers": {
                                        "type": "integer"
                                    },
                                    "steals": {
                                        "type": "integer"
                                    },
                                    "blocks": {
                                        "type": "integer"
                                    },
                                    "personal_fouls": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stat updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Stats"
                ],
                "summary": "Delete a stat",
                "description": "Delete a stat",
                "operationId": "8351a35e4c7e45cd01e4c9580e6f54b3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Stat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stat deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}